databricks_project/pyproject.toml
[tool.poetry]
name = "databricks_project"
version = "0.1"
packages = [{ include = "data_loader" }]  # your source code folder

[tool.poetry.dependencies]
python = ">3.8"
pandas = ">2.2.0"
databricks-sql-connector = ">2.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"